// libs/features/violations/src/lib/shared/enums/violation-validator-roles.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */
export enum ViolationValidatorRoles {
  COMMITEE_HEAD = 'ViolationCommitteeHead',
  COMMITEE_MEMBER = 'ViolationCommitteeMember',
  VALIDATOR_1 = 'Validator1', //First level validator in field office
  VALIDATOR_2 = 'Validator2' //Second level validator in field office
}